Add NRIC edit support
