+++ src/main/java/seedu/address/logic/commands/util/EditPersonDescriptor.java
@@ -45,45 +45,6 @@ public class EditPersonDescriptor {
     }
 
     /**
-     * Creates and returns a {@code Elderly} with the details of {@code elderlyToEdit}
-     * edited with {@code editPersonDescriptor}.
-     */
-    public static Elderly createEditedElderly(Elderly elderlyToEdit, EditPersonDescriptor editPersonDescriptor) {
-        assert elderlyToEdit != null;
-
-        Name updatedName = editPersonDescriptor.getName().orElse(elderlyToEdit.getName());
-        Phone updatedPhone = editPersonDescriptor.getPhone().orElse(elderlyToEdit.getPhone());
-        Email updatedEmail = editPersonDescriptor.getEmail().orElse(elderlyToEdit.getEmail());
-        Address updatedAddress = editPersonDescriptor.getAddress().orElse(elderlyToEdit.getAddress());
-        Nric updatedNric = editPersonDescriptor.getNric().orElse(elderlyToEdit.getNric());
-        Age updatedAge = editPersonDescriptor.getAge().orElse(elderlyToEdit.getAge());
-        Set<Tag> updatedTags = editPersonDescriptor.getTags().orElse(elderlyToEdit.getTags());
-
-        return new Elderly(updatedName, updatedPhone, updatedEmail, updatedAddress,
-                updatedNric, updatedAge, elderlyToEdit.getRiskLevel(), updatedTags);
-    }
-
-    /**
-     * Creates and returns a {@code Volunteer} with the details of {@code volunteerToEdit}
-     * edited with {@code editPersonDescriptor}.
-     */
-    public static Volunteer createEditedVolunteer(Volunteer volunteerToEdit,
-                                                  EditPersonDescriptor editPersonDescriptor) {
-        assert volunteerToEdit != null;
-
-        Name updatedName = editPersonDescriptor.getName().orElse(volunteerToEdit.getName());
-        Phone updatedPhone = editPersonDescriptor.getPhone().orElse(volunteerToEdit.getPhone());
-        Email updatedEmail = editPersonDescriptor.getEmail().orElse(volunteerToEdit.getEmail());
-        Address updatedAddress = editPersonDescriptor.getAddress().orElse(volunteerToEdit.getAddress());
-        Nric updatedNric = editPersonDescriptor.getNric().orElse(volunteerToEdit.getNric());
-        Age updatedAge = editPersonDescriptor.getAge().orElse(volunteerToEdit.getAge());
-        Set<Tag> updatedTags = editPersonDescriptor.getTags().orElse(volunteerToEdit.getTags());
-
-        return new Volunteer(updatedName, updatedPhone, updatedEmail, updatedAddress,
-                updatedNric, updatedAge, updatedTags);
-    }
-
-    /**
      * Returns true if at least one field is edited.
      */
     public boolean isAnyFieldEdited() {
